requirements.txt
numpy
pandas
pyarrow
scikit-learn
python-dotenv
pydantic
sentence-transformers
faiss-cpu
crewai
sentence-transformers
faiss-cpu
